refactor: atom, selector 분리
