+++ src/recoil/selectors.ts
@@ -1,6 +1,15 @@
 import { selector } from 'recoil';
 import { isCheckedItemIdsState, itemQuantityState, currentCartItemsState } from './atoms';
 import { Product } from '../type';
+import { fetchCartItems } from '../apis';
+
+export const cartItemsState = selector({
+  key: 'cartItemsState',
+  get: async () => {
+    const cartItems = await fetchCartItems();
+    return cartItems;
+  },
+});
 
 export const orderAmountState = selector<number>({
   key: 'orderAmountState',
